# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: FocusFlow Backend API 文件
  version: 0.1.0-phase1
  description: |
    這份文件整理 FocusFlow backend 目前 Phase-1 MVP 的 API 契約與回應口徑。

    本文件的範圍：
    - 讓 frontend、組員與老師可以快速理解目前 live backend 的 API 契約
    - 如實描述 phase-1 的 `bridge-aware` 與 runtime 可觀測回應
    - 不把 Atlas retrieval、query embedding 對齊、或 live LINE delivery 說成 production-ready

    覆蓋範圍（2026-07-24 更新）：
    - 已涵蓋 auth（含 role 驗證與 avatar）/ notifications / courses（含 PATCH/DELETE）/ videos（含 DELETE）/ watched / qa / stats / admin / LINE / student Shorts feed 的主要端點
    - 本 spec 已涵蓋主要端點，但仍非 100% 完整契約；internal processing webhook 等少數端點以 route files 為準
    - Shorts 僅代表修課 feed 與 metadata sync 已實作，不代表自動選片、剪輯、字幕或發布 worker 已完成

    本輪不處理的項目：
    - 不做 route-level JSDoc generation
    - 不做 production-hardening refactor
    - 暫不納入 internal processing webhook 的 Swagger 文件
servers:
  - url: /
    description: Same-origin backend root。Swagger UI 的 Try it out 會直接使用目前開啟的 host 與 port。
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/auth/password-reset/request:
    post:
      tags:
        - Auth
      security: []
      summary: 忘記密碼：寄送 6 位數驗證碼
      description: |
        不需登入。Email 已註冊且帳號啟用時寄出驗證碼（10 分鐘有效），同一帳號 60 秒內重複申請不重寄。
        為避免帳號探測，Email 不存在時也回相同的 200 訊息。未設定 SMTP 時回 503。
      operationId: requestPasswordReset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: 若 Email 已註冊，驗證碼已寄出
        '400':
          $ref: '#/components/responses/ValidationError'
        '502':
          description: 驗證信寄送失敗（PASSWORD_RESET_EMAIL_FAILED），驗證碼已作廢
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: 未設定寄信帳號（PASSWORD_RESET_UNAVAILABLE）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/password-reset/confirm:
    post:
      tags:
        - Auth
      security: []
      summary: 忘記密碼：以驗證碼設定新密碼
      description: 驗證碼錯誤累計 5 次即作廢；成功後驗證碼立即失效，新密碼以 bcrypt 雜湊儲存。
      operationId: confirmPasswordReset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, code, newPassword]
              properties:
                email:
                  type: string
                  format: email
                code:
                  type: string
                  pattern: '^\d{6}$'
                newPassword:
                  type: string
                  minLength: 8
      responses:
        '200':
          description: 密碼已重設
        '400':
          description: 新密碼驗證錯誤（VALIDATION_ERROR）或驗證碼無效／過期（PASSWORD_RESET_CODE_INVALID）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/me:
    get:
      tags:
        - Auth
      summary: 讀取目前登入使用者
      description: 需要有效的 bearer token。
      operationId: getCurrentUser
      responses:
        '200':
          description: 成功取得目前登入使用者
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthMeResponse'
        '401':
          description: 缺少 token、token 無效或過期，或 token 對應的使用者已不可用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                unauthorized:
                  summary: 未提供 Bearer token
                  description: 呼叫端未帶入 Bearer token，因此 backend 直接回 `UNAUTHORIZED`。
                  value:
                    success: false
                    message: Authentication required.
                    error:
                      code: UNAUTHORIZED
                invalidToken:
                  summary: Bearer token 無效或已過期
                  description: token 驗證失敗時，backend 會回 `INVALID_TOKEN`。
                  value:
                    success: false
                    message: Invalid or expired token.
                    error:
                      code: INVALID_TOKEN
        '404':
          $ref: '#/components/responses/UserNotFoundError'
    patch:
      tags:
        - Auth
      summary: 修改自己的姓名或 Email
      description: 需登入。只改姓名不需密碼；Email 為登入帳號，變更時必須提供正確的 currentPassword。Email 已被使用時回 409 DUPLICATE_RESOURCE。
      operationId: updateMyProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                  format: email
                currentPassword:
                  type: string
                  description: 變更 Email 時必填
      responses:
        '200':
          description: 已更新，回傳 data.user（公開欄位）
        '400':
          description: 驗證錯誤（VALIDATION_ERROR）或目前密碼錯誤（CURRENT_PASSWORD_INCORRECT）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '409':
          description: Email 已被其他帳號使用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/auth/me/avatar:
    get:
      tags:
        - Auth
      summary: 讀取目前登入使用者的頭貼
      operationId: getCurrentUserAvatar
      responses:
        '200':
          description: 頭貼 binary；實際 Content-Type 為 image/jpeg、image/png 或 image/webp
          headers:
            Cache-Control:
              schema:
                type: string
              example: private, max-age=0, must-revalidate
            X-Content-Type-Options:
              schema:
                type: string
              example: nosniff
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
            image/webp:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: 目前使用者沒有可讀取的頭貼（AVATAR_NOT_FOUND）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - Auth
      summary: 上傳或替換目前登入使用者的頭貼
      operationId: updateCurrentUserAvatar
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - avatar
              properties:
                avatar:
                  type: string
                  format: binary
                  description: JPEG、PNG 或 WebP，最大 5 MiB；伺服器會驗證 MIME 與檔案 signature。
      responses:
        '200':
          description: 頭貼更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AvatarUpdateResponse'
        '400':
          description: 缺檔、格式不符、multipart 錯誤或內容 signature 不符
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '409':
          description: 同一帳號的頭貼在上傳期間已被其他請求更新（AVATAR_UPDATE_CONFLICT）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: 檔案超過 5 MiB（AVATAR_TOO_LARGE）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/courses:
    get:
      tags:
        - Courses
      summary: 列出目前可存取的課程
      description: |
        依照目前登入者的角色與存取規則，回傳可見的課程列表。

        student 僅回傳 `active Enrollment ∩ Course.status=published`；公開狀態本身不是學生存取授權。teacher / admin 依各自管理範圍回傳。

        這個回應在 Phase-1 是 `bridge-aware`，可能會帶出 `qaScopeOnly`、`bridgeMode` 等欄位，讓 frontend 或 demo 使用者可以區分一般課程與 QA-only 的 bridge 呈現。
      operationId: listCourses
      responses:
        '200':
          description: 成功取得課程列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseListResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    post:
      tags:
        - Courses
      summary: 建立課程
      description: 僅限 teacher / admin 使用的課程建立端點。
      operationId: createCourse
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseCreateRequest'
      responses:
        '201':
          description: 課程建立成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseItemResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /api/v1/courses/{courseId}:
    get:
      tags:
        - Courses
      summary: 依 courseId 讀取單一課程
      description: |
        回傳目前使用者可存取的單一課程。

        這個回應在 Phase-1 仍是 `bridge-aware`，可能帶出 `qaScopeOnly`、`bridgeMode` 與 bridge contract 提示欄位。
      operationId: getCourseById
      parameters:
        - $ref: '#/components/parameters/CourseId'
      responses:
        '200':
          description: 成功取得課程
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseItemResponse'
        '400':
          $ref: '#/components/responses/InvalidIdError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/CourseAccessDeniedError'
        '404':
          $ref: '#/components/responses/CourseNotFoundError'
    patch:
      tags:
        - Courses
      summary: 部分更新課程
      description: 僅限 teacher（owner）/ admin 使用。可更新標題、描述、狀態等課程欄位。
      operationId: updateCourse
      parameters:
        - $ref: '#/components/parameters/CourseId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseCreateRequest'
      responses:
        '200':
          description: 課程更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseItemResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/CourseManageDeniedError'
        '404':
          $ref: '#/components/responses/CourseNotFoundError'
    delete:
      tags:
        - Courses
      summary: 刪除課程（cascade）
      description: |
        僅限 admin 或該課程的 owner teacher 使用。

        刪除會 cascade 清理對應的 Video / Segment / transcripts、`course.videoIds`、`Enrollment` 與 `User.activeCourseId`；
        但**保留** UsageLog / Question 歷史紀錄（改由 display 層分流顯示「內容已下架」）。
      operationId: deleteCourse
      parameters:
        - $ref: '#/components/parameters/CourseId'
      responses:
        '200':
          description: 課程刪除成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessBase'
                  - type: object
                    properties:
                      data:
                        description: cascade 刪除結果摘要（清理的關聯資源數量等）。
        '400':
          $ref: '#/components/responses/InvalidIdError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/CourseManageDeniedError'
        '404':
          $ref: '#/components/responses/CourseNotFoundError'
...
  /api/v1/courses/{courseId}/videos/{videoId}/watched:
    post:
      tags:
        - Videos
      summary: 標記影片為已觀看完成
      description: |
        學生用端點。驗證學生身分與「影片屬於該課程」後，`$addToSet` 寫入 `Enrollment.watchedVideoIds`，
        並重算 `progress = watched / total × 100`。第一次觀看時會額外寫一筆 `UsageLog event=WATCH`（重複觀看不重複寫）。
      operationId: markVideoWatched
      parameters:
        - $ref: '#/components/parameters/CourseId'
        - $ref: '#/components/parameters/VideoId'
      responses:
        '200':
          description: 已標記觀看完成
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessBase'
                  - type: object
                    properties:
                      data:
                        description: 更新後的 enrollment 進度摘要（watchedVideoIds 數量、progress 百分比等）。
        '400':
          $ref: '#/components/responses/InvalidIdError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/CourseAccessDeniedError'
        '404':
          $ref: '#/components/responses/VideoNotFoundError'
  /api/v1/courses/{courseId}/videos/{videoId}/opened:
    post:
      tags:
        - Videos
      summary: 記錄學生點開影片
      description: |
        學生用端點，驗證規則同 `/watched`（active enrolled student、影片屬於該課程）。每次點開都寫一筆
        `UsageLog event=video_open` 供管理員統計；不改 `Enrollment.watchedVideoIds` 與 `progress`。
      operationId: markVideoOpened
      parameters:
        - $ref: '#/components/parameters/CourseId'
        - $ref: '#/components/parameters/VideoId'
      responses:
        '200':
          description: 已記錄點開
        '400':
          $ref: '#/components/responses/InvalidIdError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/CourseAccessDeniedError'
        '404':
          $ref: '#/components/responses/VideoNotFoundError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    LineSignature:
      type: apiKey
      in: header
      name: x-line-signature
      description: |
        使用 `LINE_CHANNEL_SECRET`，對「完整且未改動的 raw webhook body」計算出的 HMAC-SHA256 signature。

        這是一個 `integration-facing` 的簽章驗證流程；Swagger UI 不會自動替你計算 signature。
  parameters:
    CourseId:
      name: courseId
      in: path
      required: true
      description: 目標課程的 MongoDB ObjectId。
      schema:
        $ref: '#/components/schemas/ObjectId'
    VideoId:
      name: videoId
      in: path
      required: true
      description: 目標影片的 MongoDB ObjectId。
      schema:
        $ref: '#/components/schemas/ObjectId'
...
  responses:
    ValidationError:
      description: request payload 驗證失敗。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            validation:
              summary: 一般欄位驗證失敗
              description: 代表 request 結構存在欄位缺漏、型別錯誤，或其他基本驗證失敗情況。
              value:
                success: false
                message: Validation failed.
                error:
                  code: VALIDATION_ERROR
                  details:
                    - title is required
...
    UnauthorizedError:
      description: 缺少 bearer token，或 token 對應的使用者目前不可用。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            unauthorized:
              summary: 尚未通過身份驗證
              description: 請求未帶入必要的 Bearer token，或目前請求情境需要先完成認證。
              value:
                success: false
                message: Authentication required.
                error:
                  code: UNAUTHORIZED
    ForbiddenError:
      description: 已登入，但角色不符合此端點要求。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            forbidden:
              summary: 角色權限不足
              description: 呼叫者已登入，但角色不符合這個端點的授權條件。
              value:
                success: false
                message: You do not have permission to perform this action.
                error:
                  code: FORBIDDEN
    InvalidIdError:
      description: path 或 body 內的 ObjectId 格式不合法。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            invalidId:
              summary: ObjectId 格式錯誤
              description: path 或 body 中的資源識別值不是合法的 MongoDB ObjectId。
              value:
                success: false
                message: Invalid resource id.
                error:
                  code: INVALID_ID
    UserNotFoundError:
      description: token 驗證成功，但對應的使用者已不存在。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            userNotFound:
              summary: 找不到 token 對應的使用者
              description: token 本身可被驗證，但資料庫中已找不到對應 user。
              value:
                success: false
                message: User not found.
                error:
                  code: USER_NOT_FOUND
    CourseAccessDeniedError:
      description: 呼叫端沒有存取這門課的權限。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            courseAccessDenied:
              summary: 沒有課程存取權
              description: 呼叫者雖已登入，但不符合這門課目前的課程存取規則。
              value:
                success: false
                message: You do not have access to this course.
                error:
                  code: COURSE_ACCESS_DENIED
    CourseManageDeniedError:
      description: 呼叫端沒有管理這門課或影片的權限。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            courseManageDenied:
              summary: 沒有課程管理權
              description: 呼叫者不是課程 owner teacher，也不是 admin，因此不能執行管理型操作。
              value:
                success: false
                message: You do not have permission to manage this course.
                error:
                  code: COURSE_MANAGE_DENIED
    CourseNotFoundError:
      description: 找不到指定課程。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            courseNotFound:
              summary: 課程不存在
              description: 請求中的課程識別值格式正確，但資料庫中找不到對應課程。
              value:
                success: false
                message: Course not found.
                error:
                  code: COURSE_NOT_FOUND
    VideoNotFoundError:
      description: 找不到指定影片。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            videoNotFound:
              summary: 影片不存在
              description: 請求中的影片識別值格式正確，但資料庫中找不到對應影片。
              value:
                success: false
                message: Video not found.
                error:
                  code: VIDEO_NOT_FOUND
...
  schemas:
    ObjectId:
      description: MongoDB ObjectId 的字串格式。
      type: string
      pattern: '^[a-fA-F0-9]{24}$'
      example: 507f191e810c19729de860eb
    UserRole:
      description: 系統目前支援的使用者角色列舉。
      type: string
      enum:
        - admin
        - teacher
        - student
    CourseStatus:
      description: 課程狀態列舉。
      type: string
      enum:
        - draft
        - published
        - archived
    BridgeMode:
      description: |
        Phase-1 課程或影片的 `bridge` 呈現模式。

        `bridge` 代表目前資料仍處於 app-owned records 與 pipeline metadata 共存的過渡狀態。
        `qa_scope_only` 代表這份資料主要只用來提供 QA 範圍，不保證具備完整 app-owned video 行為。
      type: string
      enum:
        - standard
        - qa_scope_only
        - mixed_scope
...
    SuccessBase:
      description: 所有成功回應共用的 envelope 基底。
      type: object
      required:
        - success
        - message
      properties:
        success:
          const: true
        message:
          type: string
    ErrorResponse:
      description: 所有錯誤回應共用的 envelope 格式。
      type: object
      required:
        - success
        - message
        - error
      properties:
        success:
          const: false
        message:
          type: string
        error:
          type: object
          required:
            - code
          properties:
            code:
              type: string
            details:
              description: 可選的 endpoint-specific error details；實際結構會依 error 類型而不同。
...
    PublicUser:
      description: 對 API 使用者公開的 user 資訊，不包含敏感欄位。
      type: object
      required:
        - id
        - name
        - email
        - role
        - isActive
        - hasAvatar
        - avatarUpdatedAt
        - isLineBound
        - lineBindAt
        - activeCourseId
        - lineConversationState
      properties:
        id:
          $ref: '#/components/schemas/ObjectId'
        name:
          type: string
        email:
          type: string
          format: email
        role:
          $ref: '#/components/schemas/UserRole'
        isActive:
          type: boolean
        hasAvatar:
          description: 目前帳號是否有可由 authenticated avatar API 讀取的頭貼。
          type: boolean
        avatarUpdatedAt:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
        isLineBound:
          description: 是否已綁定 LINE 帳號；不暴露實際 LINE user id。
          type: boolean
        lineBindAt:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
        activeCourseId:
          anyOf:
            - $ref: '#/components/schemas/ObjectId'
            - type: 'null'
        lineConversationState:
          type: string
          enum:
            - idle
            - awaiting_course_selection
        createdAt:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
        updatedAt:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
    TeacherSummary:
      description: 課程或影片回應中使用的 teacher 簡化資料。
      type: object
      required:
        - _id
        - name
        - email
        - role
        - isActive
      properties:
        _id:
          $ref: '#/components/schemas/ObjectId'
        name:
          type: string
        email:
          type: string
          format: email
        role:
          $ref: '#/components/schemas/UserRole'
        isActive:
          type: boolean
    CourseBridgeFields:
      description: |
        課程在 Phase-1 下的 `bridge` 補充欄位。

        這些欄位用來說明目前課程是一般課程、混合課程，或僅作為 QA scope 的 bridge 課程。
      type: object
      required:
        - isBridgeCourse
        - qaScopeOnly
        - bridgeMode
        - videoCount
        - appVideoCount
        - bridgeVideoCount
        - appOwnedVideoCount
        - metadataOnlyVideoCount
        - courseVideoRefCount
        - bridgeExternalVideoIds
        - courseVideoIds
        - bridgeContract
        - bridgeContractPath
      properties:
        isBridgeCourse:
          description: "`true` 代表這門課目前不是純 `standard` 呈現，而是 bridge-first 的課程視圖。"
          type: boolean
        qaScopeOnly:
          description: "`true` 代表這門課目前偏向 QA scope 使用，而不是完整 app-owned course 體驗。"
          type: boolean
        bridgeMode:
          $ref: '#/components/schemas/BridgeMode'
        videoCount:
          type: integer
          minimum: 0
        appVideoCount:
          description: 課程下 app-owned video 的正式統計欄位。
          type: integer
          minimum: 0
        bridgeVideoCount:
          description: 課程下 bridge / pipeline metadata video 的正式統計欄位。
          type: integer
          minimum: 0
        appOwnedVideoCount:
          description: "`appVideoCount` 的 readability alias，統計來源完全相同，僅用於讓 frontend / demo 更容易理解。"
          type: integer
          minimum: 0
        metadataOnlyVideoCount:
          description: "`bridgeVideoCount` 的 readability alias，統計來源完全相同，僅用於強調這些影片目前是 `metadataOnly` bridge video。"
          type: integer
          minimum: 0
        courseVideoRefCount:
          type: integer
          minimum: 0
        bridgeExternalVideoIds:
          type: array
          items:
            type: string
        courseVideoIds:
          type: array
          items:
            $ref: '#/components/schemas/ObjectId'
        bridgeContract:
          description: backend 目前採用的 bridge contract 名稱。
          type: string
          example: course_video_refs_v1
        bridgeContractPath:
          description: bridge contract 的資料串接路徑說明。
          type: string
          example: course.videoIds -> videos._id -> videos.videoId -> video_segments_text.videoId
    Course:
      description: 課程主體回應，含基本課程資料與 Phase-1 bridge 欄位。
      allOf:
        - type: object
          required:
            - _id
            - title
            - description
            - teacherId
            - videoIds
            - status
          properties:
            _id:
              $ref: '#/components/schemas/ObjectId'
            title:
              type: string
            description:
              type: string
            teacherId:
              $ref: '#/components/schemas/TeacherSummary'
            videoIds:
              type: array
              items:
                $ref: '#/components/schemas/ObjectId'
            status:
              $ref: '#/components/schemas/CourseStatus'
            createdAt:
              anyOf:
                - type: string
                  format: date-time
                - type: 'null'
            updatedAt:
              anyOf:
                - type: string
                  format: date-time
                - type: 'null'
        - $ref: '#/components/schemas/CourseBridgeFields'
...
    AvatarUpdateResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessBase'
        - type: object
          required:
            - data
          properties:
            data:
              type: object
              required:
                - user
                - avatar
              properties:
                user:
                  $ref: '#/components/schemas/PublicUser'
                avatar:
                  type: object
                  required:
                    - mimeType
                    - updatedAt
                  properties:
                    mimeType:
                      type: string
                      enum:
                        - image/jpeg
                        - image/png
                        - image/webp
                    updatedAt:
                      type: string
                      format: date-time
...
    AuthMeData:
      description: 目前登入使用者查詢的資料主體。
      type: object
      required:
        - user
      properties:
        user:
          $ref: '#/components/schemas/PublicUser'
    CourseCreateRequest:
      description: 建立課程時使用的 request body。
      type: object
      required:
        - title
      properties:
        title:
          type: string
        description:
          type: string
        status:
          $ref: '#/components/schemas/CourseStatus'
        teacherId:
          $ref: '#/components/schemas/ObjectId'
...
    AuthMeResponse:
      description: "`/api/v1/auth/me` 的成功回應 envelope。"
      allOf:
        - $ref: '#/components/schemas/SuccessBase'
        - type: object
          required:
            - data
          properties:
            data:
              $ref: '#/components/schemas/AuthMeData'
    CourseListResponse:
      description: 課程列表端點的成功回應 envelope。
      allOf:
        - $ref: '#/components/schemas/SuccessBase'
        - type: object
          required:
            - data
          properties:
            data:
              type: object
              required:
                - courses
              properties:
                courses:
                  type: array
                  items:
                    $ref: '#/components/schemas/Course'
    CourseItemResponse:
      description: 單一課程端點的成功回應 envelope。
      allOf:
        - $ref: '#/components/schemas/SuccessBase'
        - type: object
          required:
            - data
          properties:
            data:
              type: object
              required:
                - course
              properties:
                course:
                  $ref: '#/components/schemas/Course'
...